Add directory for bash scripts with very limited use cases
In the last commit changing the vim plugins to proper submodules,
I created two supporting scripts, which might come in handy,
if there are going to be additional vim plugins for which the
procedure has to be repeated.

There was no proper folder for these scripts so I added one.