Add wrapping function for git; Add `git root` function
There is no possibility to run a function in the same shell,
if defining an alias in .gitconfig. They all run in subshells.
This is supoptimal if your function using `cd` for example.
Bash functions however have a higher priority if they are self-defined.
Defining a custom wrapper function with the same name works,
as the function gets called instead of the command.
MAKE SURE to use the command keyword in front of the wrapped function
to avoid a recursive endless calling loop.
First example using htis method is `git root`, which can now be called
like a normal git command. (with a whitespace between git and root)