Update start alias to use wslview
Wslview is part of the [wslu](https://github.com/wslutilities/wslu)
package, which comes preinstalled with the ubuntu-wsl version.
After upgrading to wsl2, the previous version relying on cmd.exe,
failed, due to the new way of accessing wsl files. They are now stored
under a network directory at `\\$wsl\...`, but cmd.exe can  not work
with that path. It might be possible to adapt the previous version but
as wslview comes preinstalled, it's not worth the hassle.