Remove fzf from apt; Rely on version installed for vim
The fzf version on the package index of ubuntu apt, is outdated. Current
version is 0.24, indexed version is 0.20. Therefore we rely on a version
that is automatically installed if we want to use fzf for vim too. The
plug plugin manager clones the git repository for fzf anyways, so we can
rely on that version. Adding that repository as a submodule and removing
the previous version of fzf including key-bindings.