Change to use default existing subdirectory path or passed argument
Updated the implementation of the submodule adder, so that one can pass
in the parent directory of the existing submodule folders.
The plug manager for vim stores all plugins in a directory
"vim/.vim/plugged", meaning it clones all repositories into folders
under that parent directory. Previous iteration was meant to be called
from the dotfile base folder, while this one now checks the passed in
path for the existing repos.