Set proper file viewers for vifm.
There are infinite options here.
Most people recommend ueberzug and vifmimg for image previews.
This however does not work in WSL. The X11 server is missing as Windows
uses a different window manager.
Hence ueberzug and most image previewers do not work.
There is an issue on the Windows Terminal github, for an option of
graphical image displays, but there is no solution for that currently.

I tried shellpick as an aternative, with the hope that it works as it is
terminal-only, but it does fail too. Some error related to tty, which
most likely is also related to WSL. Abandoned the topic for now, and
replaced all image previews with `magick identify`.