Change parser directory to `~/.local/share`;
This is the directory where tree-sitter is looking for grammar files
when highlighting files.

"""
tree-sitter highlight command takes one or more file paths, and tries to
automatically determine which language should be used to highlight those
files. In order to do this, it needs to know where to look for
Tree-sitter grammars on your filesystem. You can control this using the
"parser-directories" key in your configuration file:

```json
{
  "parser-directories": [
    "/Users/my-name/code",
    "/Users/my-name/other-code"
  ]
}
```

Currently, any folder within one of these parser directories
whose name begins with “tree-sitter-“ will be treated as a
Tree-sitter grammar repository.
"""