Link used colorschemes into colors folder
Prior to this change I had a setup script to put the vim-colorscheme
into the correct folder.
Now the currently used files are symlinked, so that upon deletion of the
used programs, they disappear too.
And the naming gets easier.
Truoble now is updating the used colorscheme, as it needs to be
symlinked into the color folder every time it's changed.
I'm thinking about making a symlink inside the programs colorscheme
folder for the currently used scheme and symlink that into my colors
folder, so that I only need to update the symlink from inside the
colorscheme folder, instead of the full symlinks into my colors folder.