Set tmux default terminal to xterm-256
TMUX sets its $TERM as screen. So, no 256 color support and some other
differences. I noted these as I was opening vim, and the background of
the line number was set differently.
As a quick fix, I set TMUX default terminal to xterm-256, which is the
same as my default terminal emulator.
However setting $TERM inside the terminal is often regarded as a
mistake, since the same config can then not be used across multiple
different terminal emulators. Since it is "hardcoded" which temrinal
emulator features are supported.

There might be a better way to handle this.
I could adapt my vim colorscheme, by also setting the guifg and guibg.
At least, thats what I think. But for now, it works and *shrugs*.