Insert environment check before attempting to source cargo
The changed line currently sources the environment for cargo.
This should only be done when the file is actually present.
There are however better checks. One could check if cargo is installed
in the first place.
Second, do not look for a hardcoded environment, but instead use the
cargo and rust environment variables, $RUST_HOME and $CARGO_HOME I
believe.