Guard <TAB> against missing UltiSnippets
Prior to this change, the use of UltiSnips expansion function was
hard-coded into my tab-handling. With the option of UltiSnips now
missing in my Vim setup, it had to be guarded against.
I created a function, that handles snippet expansion and jumping.
It returns 0/false if ultisnips is not present, else it returns the
proper value of jump or expansion.

<Tab> now inserts a tab even if pressed in the middle of a word.