Source environment variables set in pre_deploy script
Several machine-specific env vars are set in my deployment script.
Sourcing is the only way to get them to work currently.
There is however a PR open on github/dotter which would allow setting
env vars via the normal pre_deploy.sh hook instead of customizing it in
my way.

It references the filled-in template in the dotter cache.