Add completion to <TAB> key
The additional empty string return is due to vimscript automatically
returning the last statement if no return is given. Triggering
completion always exited with a success code of 0.( Another Vimscript
oddity) This lead to the "0" always being inserted into the buffer and
screwing completion, as most completions do not end with a "0", hence no
completion option was found.