Disable ALE's LSP features
ALE can function as a LSP client itself.
There is no need for that, as I am using the builtin Neovim version.
So, to avoid running the LSP twice for each buffers content ALE does not
run the LSP.

Later, I intended to send the LSP diagnostics to ALE, so all
diagnostics, whether LSP or other tools are at the same location.