Change nvim clipboard behaviour to not auto-copy to system clipboard
Previously any copying inside nvim automatically put the yanked text
into the system clipboard. Now it is only copied to system clipboard
upon yanking to the '*' or '+' register.
Available options for the yanking behaviour are 'unnamed' which defaults
to yanking in the '*' register and 'unnamedplus' which yanks into the
'+' register.