Adjust Tab-mapping to breaking change in nvim-cmp
With PR https://github.com/hrsh7th/nvim-cmp/pull/224, nvim-cmp uses
floating windows instead of the vim-internal pummenu. The previous
function to resolve the <Tab> key relied on the `pumvisible` function to
check for an open completion window, which is not possible anymore.
The new function features an extra branch checking for an open
completion window, via `cmp.visible()`..