Add mappings allowing for window switch and normal mode in vim term
Vim terminal mode is a special version of insert mode in a vim terminal.
It does not allow for vim editing etc. The default way out of that
version of insert mode  is `<C-\><C-n>`, which is the normal way to
directly jump to normal mode.

'\' does however not work properly, with my keyboard setup.
The combination '<alt-ctrl-ß>', which normally prints a backslash, does
not work when remapping the '<ctrl>' in the same command.

Per default, Stylua deletes '\' in strings unless they escape something
with special meaning, hence it needed to be disabled for this certain
section.