Change UI toggle to simple open
Instead of toggling the debugging UI on keypress, just try to open it.
This works better, since it allows for easier rerunning of debugging and
does not allow the toggling and debugging to get out of sync, e.g.
opening the UI while not debugging and closibg the UI when starting to
debug.