Link $BROWSER to wsl-aware opening program
`rustup doc` relies on both xdg-open and $BROWSER. Both have to be set
in order to use the command. I had both properly set.
$BROWSER was set to the windows instance of firefox, however quoting
issues made it unusable.
xdg-open was aliased to wslview.
Neither transforms the path of the passed arguments, hence this failed
with the rustup doc command as it was referencing a file using the unix
way instead of the adulterated wsl path.

Fixed now but I dont like this fix as BROWSER is set improperly.
It does not open internet adresses, but only on-disk html files.