fix(nvim): extract web search capability from diagnostic search
This commit creates a separate 'lua/utils/websearch.lua' file, that
contains the web search functionality.
This way, the functionality can be used separately in the future.

The configuration is passed-through from the diagnostic search and a
configured instance is stored in the diagnostic search module.
*Dependency injection*

Part of this is also a new function which uses vim.ui.input to get the
input for a web search.

Also uses the more modern 'vim.system'.
And makes the config parameter optional.