feat(tmux): bind ? to fzf-based key listing
This is displaying a new command for listing the keys of TMUX hotkeys.
It is displayed in a popup, because that's an interactive session and as
such can run an interactive program like fzf.

For some reason, sourcing the script from the fzf directory does not
work despite being exactly the same command as the hardcoded one.